Get java classpath

Some cluster configuration packages contain java dependencies. This utility will use the jar names from the configuration YAML and find the full file paths to the stated dependencies.

get_java_classpath(jars, cluster_name, base_path, configs)

get_java_dependencies_path(cluster_name, pkg_name)

Arguments

jars
Character. A vector of jar files to include on the classpath. These should be listed using the full file name, including the ".jar" extension at the end.

cluster_name
Character. The name of the cluster. Capitalization and spacing is ignored and does not matter (e.g., "My Cluster" is equivalent to "mycluster")

base_path
Character. A file path to the directory where the dependencies may be found. This is used in place of the cluster resources if provided.

configs
List. The parsed YAML configurations. This is used if provided and neither cluster_name nor base_path is provided